started main part of settings and loading overhaul

+++ Hecaera - Awakened/Assets/Scripts/Game Saving/SettingsSaveData.cs
@@ -5,12 +5,12 @@ using UnityEngine;
 public class SettingsSaveData
 {
     [Header("Sensitivity")]
-    public float horizontalSensitivity = 0f; //Goes between -100 and 100, -100 meaning half the sensitivity of the default. 100 meaning default + half
-    public float verticalSensitivity = 0f; //Goes between -100 and 100
+    public int horizontalSensitivity = 0; //Goes between -100 and 100, -100 meaning half the sensitivity of the default. 100 meaning default + half
+    public int verticalSensitivity = 0; //Goes between -100 and 100
 
     [Header("Volumes")]
-    public float mainVolume = 50f;
-    public float musicVolume = 25f;
-    public float SFXVolume = 50f;
-    public float dialogVolume = 50f;
+    public int mainVolume = 50;
+    public int musicVolume = 25;
+    public int SFXVolume = 50;
+    public int dialogVolume = 50;
 }
